+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     - Fix parsing if the fourth part of the user agent engine version (split by `.`) starts with `0[0-9]` ([#35](https://github.com/elixir-inspector/ua_inspector/issues/35))
 
 - Enhancements
+    - Default upstream database version is now `6.3.1`
     - Upstream improvements for "360 Secure" browser detection have been integrated
     - Upstream improvements for "Arachne" browser engine detection have been integrated
     - Upstream improvements for "Clecko" browser engine detection have been integrated